Move SurfaceContainer to package
