+++ core/src/test/java/com/novoda/noplayer/internal/exoplayer/ExoPlayerTwoImplTest.java
@@ -601,7 +601,7 @@ public class ExoPlayerTwoImplTest {
         @Mock
         View containerView;
         @Mock
-        PlayerView.SurfaceContainer surfaceContainer;
+        SurfaceContainer surfaceContainer;
 
         ExoPlayerTwoImpl player;
